refactor(auth): merge AuthProvider and useAuth into AuthContext
- Combined AuthProvider and useAuth into a single AuthContext file to simplify context management and reduce file complexity.
- Updated signUp and signIn to use destructured parameters.
- Adjusted App.jsx to import from the merged file.
