--- src/context/AuthContext.js
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const AuthContext = createContext();
