--- src/context/useAuth.js
@@ -1,6 +1,0 @@
-import  { useContext } from "react";
-import { AuthContext } from "src/context/AuthContext";
-
-export function useAuth() {
-  return useContext(AuthContext);
-}
